assignment 3/pyproject.toml
[project]
name = "assignment-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "google-generativeai>=0.8.5",
]
